+++ sentry-spring/src/main/java/io/sentry/spring/tracing/SentryTracingFilter.java
@@ -73,6 +73,11 @@ public class SentryTracingFilter extends OncePerRequestFilter {
     if (hub.isEnabled()) {
       final String sentryTraceHeader = httpRequest.getHeader(SentryTraceHeader.SENTRY_TRACE_HEADER);
 
+      hub.configureScope(
+          scope -> {
+            scope.setRequest(requestResolver.resolveSentryRequest(httpRequest));
+          });
+
       // at this stage we are not able to get real transaction name
       final ITransaction transaction = startTransaction(httpRequest, sentryTraceHeader);
       try {
@@ -86,7 +91,6 @@ public class SentryTracingFilter extends OncePerRequestFilter {
         if (transactionName != null) {
           transaction.setName(transactionName);
           transaction.setOperation(TRANSACTION_OP);
-          transaction.setRequest(requestResolver.resolveSentryRequest(httpRequest));
           transaction.setStatus(SpanStatus.fromHttpStatusCode(httpResponse.getStatus()));
           transaction.finish();
         }
@@ -109,9 +113,7 @@ public class SentryTracingFilter extends OncePerRequestFilter {
         final TransactionContext contexts =
             TransactionContext.fromSentryTrace(
                 name, "http.server", new SentryTraceHeader(sentryTraceHeader));
-        final ITransaction transaction = hub.startTransaction(contexts, customSamplingContext);
-        hub.configureScope(scope -> scope.setTransaction(transaction));
-        return transaction;
+        return hub.startTransaction(contexts, customSamplingContext);
       } catch (InvalidSentryTraceHeaderException e) {
         hub.getOptions()
             .getLogger()