+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+* Feat: Add an overload for `startTransaction` that sets the created transaction to the Scope #1313
+* Ref: Separate user facing and protocol classes in the Performance feature (#1304)
+
 # 4.3.0
 
 * Fix: Fix setting in-app-includes from external properties (#1291)